Fix #19 DataUriEncodingTest.testDataUriEncode fails on Java 11
When reading/write an image, the behavior may (will!) different
from one Java version to another.

No longer use any image API in this test.
Rational : The purpose of the method Utils.getDataUriForImage being
tested is not to read or write an image by following some image format
specification. But simply to ensure that a specified byte array
(representing an image) gets written using the Data URI specification
in base64.

Rewrite the test just to check that :
- remove all image loading/writing
- simply check the output match the expected base 64 data uri syntax.